+++ packages/sitecore-jss-nextjs/src/context/context.test.ts
@@ -171,15 +171,16 @@ describe('Context', () => {
       expect(context.siteName).to.equal('website-1');
     });
 
-    it('should catch and log when SDK initialization rejects', () => {
-      const consoleSpy = sinon.spy(console, 'log');
+    it('should not fail context init when an SDK init throws', (done) => {
       const localProps = { ...props, sdks: errorSdk };
       const context = new Context<typeof errorSdk>(localProps);
-      context.init();
-      expect(
-        consoleSpy.calledWith('Initialization for SDK Error skipped. Reason: \n Cannot init Error')
-      );
-      consoleSpy.restore();
+      try {
+        context.init();
+      } catch (e) {
+        done(e);
+      } finally {
+        done();
+      }
     });
 
     it('should reject when getting SDK that rejected initialization', (done) => {
@@ -189,7 +190,7 @@ describe('Context', () => {
       context
         .getSDK('Error')
         .then(() => {
-          throw new Error('should not resolve');
+          done('should not resolve');
         })
         .catch((e) => {
           expect(e).to.be.equal('Cannot init Error');